feat: updated tooltip design
